Fix type of early return variable mapped to another variable in the region's constant table
A early return variable mapped to either a constant or another variable in the region's constantTable will have the type `AstVarExpr`. But, SpfToGreenVisitor used to replace it with its corresponding Green variable type sibling if the early return variable was found to not be mapped to a constant. This commit fixes this mistake and allows the early return variable AstVarExpr type to be preserved so that a lookup later in the constantTable in VeritestingListener.pushExpOnStack does not fail. Also, allowing the lookup in VeritestingListener.pushExpOnStack to check if the early return variable is a variable or a constant in the region's constantTable.

Testing:
This commit fixes the issues we were seeing in the list1.java example from the jbmc-regression test suite.