Alpha-renaming in the constants table should be on both keys and values
Alpha-renaming in the constantsTable in the region was previously done only on the keys. But it can be done both on the keys and values. Also, we forgot to alpha-rename AstVarExpr variables (which are currently only used for summarizing early returns).

This fixes the equivalence-checking issues I was seeing with the replace benchmark.