Small-values concretization for symbolic array lengths
Turning off bounded value concretization and instead implementing small-values concretization for symbolic array lengths. This optimization helps find the right answer for InsertionSort-FunUnsat01 and MergeSortIterative-MemUnsat01. At the same time, it also causes SPF to explore further in the BellmanFord-FunUnsat01 and BellmanFord-MemUnsat01 benchmarks to find array-indexed-out-of-bounds accesses in these two benchmarks.