Adding small values concretization for multinewarray bytecode insn
This feature is implemented in a way similar to the small values concretization done for symarrays/newarray. This feature is required for SV-COMP 2020 benchmarks like algorithms/Tsp-FunSat01, Tsp-FunUnsat01, Tsp-MemSat01, Tsp-MemUnsat01, BellmanFord-MemUnsat02, BellmanFord-MemSat02, BellmanFord-FunUnsat02, BellmanFord-FunSat02. I was able to correctly implement this feature. But, SPF later crashes on finding an ArrayExpression object being typecast to a SymbolicInteger in bytecode/GETFIELD.java#187. I am guessing this is due to a bad interaction between the symbolic arrays feature and the heap path condition.