Register a `ClassInfo` before using it to get `ElementInfo` from it
Since JR summarizes code without executing it, it can run into situations where the `ClassInfo` it is trying to use has not been initialized yet. One way to address this is for Ranger to do the initialization. This fix helps create and link the `ElementInfo` to the `ClassInfo` that can subsequently be used.